fix(voice): revert to plan's float+subtract wrap (was unbounded double+fmod)
Also fixes test_naive_saw_wraps to use 1500 Hz (phase_inc=1/32 exact in
float) so the wrap test doesn't rely on 48*(1/48) accumulating to exactly
1.0 in float — it doesn't.

+++ src/voice.h
@@ -11,8 +11,8 @@ class Voice {
 
   private:
     float sample_rate_ = 48000.0f;
-    double phase_      = 0.0;      // [0, 1) - double precision to reduce accumulation error
-    double phase_inc_  = 0.0;      // freq / sample_rate
+    float phase_       = 0.0f;     // [0, 1)
+    float phase_inc_   = 0.0f;     // freq / sample_rate
 };
 
 }  // namespace sawstack
